requirements.txt
# Core web dependencies (version-locked for compatibility)
fastapi==0.104.1
python-multipart==0.0.6
pydantic>=2.4.2
uvicorn[standard]==0.24.0
aiofiles==23.2.1
requests==2.31.0

# Speech synthesis
gTTS==2.4.0

# For local LLM inference (GGUF format):
ctransformers>=0.2.24  # Efficient inference for GGUF models

# For local speech recognition:
openai-whisper>=20231106

# Optional / advanced (uncomment if needed):
# bitsandbytes  # for 4-bit/8-bit quantization
# TTS  # for local Coqui TTS instead of gTTS
